requirements.txt
streamlit
openai
python-dotenv
sentence-transformers
torch
transformers
tqdm
scikit-learn
numpy
